fix: detail페이지 vars 모듈 불러오는 방법을 수정했습니다.
